% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int.R
\name{interval_estimated}
\alias{interval_estimated}
\title{Interval estimation using R}
\usage{
interval_estimated(x, sigma, side, alpha)
}
\arguments{
\item{x}{the inpput data}

\item{sigma}{the overall standard deviation}

\item{side}{one side or two side interval estimation}

\item{alpha}{cofidence}
}
\value{
the interval
}
\description{
One side or two side interval estimation in any confidence using R
}
\examples{
\dontrun{
sigma=-1
side=0
alpha=0.05
x= c(28,26,33,24,34,-44,27,16,40,-2,29,22,24,21,25,30,23,29,31,19)
interval_estimated(x,sigma,side,alpha)
}
}
