% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min.R
\name{min}
\alias{min}
\title{Take the minimum using R}
\usage{
min(x)
}
\arguments{
\item{X}{the input sequence}
}
\value{
the minimum of X
}
\description{
Take the minimum using R
}
\examples{
\dontrun{
x=c(311,365,343,377,254)
y=min(x)
print(y)
}
}
