% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tes.R
\name{tes}
\alias{tes}
\title{A two-sample Cramer-von Mises test using R}
\usage{
tes(x, y)
}
\arguments{
\item{X}{The data under an certain distribution}

\item{Y}{The data under an certain distribution}
}
\value{
the test chart
}
\description{
Implement the two-sample Cramer-von Mises test for equal distributions as a permutation test.
}
\examples{
\dontrun{
attach(chickwts)
x <- sort(as.vector(weight[feed == "soybean"]))
y <- sort(as.vector(weight[feed == "linseed"]))
}
}
